Bin a GRanges.

BinGRanges Bin a GRanges and allow to apply a summary method (e.g: 'mean', 'median', 'sum', 'max, 'min' ...) to a chossen numericals variables of ranges in a same bin.

Usage

BinGRanges(
  gRange.gnr = NULL,
  chromSize.dtf = NULL,
  binSize.int = NULL,
  method.chr = "mean",
  variablesName.chr_vec = NULL,
  na.rm = TRUE,
  cores.num = 1,
  reduce.bln = TRUE,
  verbose.bln = TRUE
)

Arguments

gRange.gnr

<GRanges>: a GRanges to bin.

chromSize.dtf

<data.frame>: a data.frame where first colum correspond to the chromosomes names, and the second column correspond to the chromosomes lengths in base pairs.

binSize.int

<integer>: width of the bins.

method.chr

<character>: name of a summary method as 'mean', 'median', 'sum', 'max, 'min'. (Default 'mean')

variablesName.chr_vec

<character> : a character vector that specify the metadata columns of GRanges on which apply the summary method.

na.rm

<logical> : a logical value indicating whether 'NA' values should be stripped before the computation proceeds. (Default TRUE)

cores.num

<integer> : the number of cores. (Default 1)

reduce.bln

<logical> : whether duplicated Bin must been reduced with de summary method. (Default TRUE)

verbose.bln

<logical>: if TRUE show the progression in console. (Default TRUE)

Value

A binned GRanges.

Examples

GRange.gnr <- GenomicRanges::GRanges(
    seqnames = S4Vectors::Rle(c("chr1", "chr2"), c(3, 1)),
    ranges = IRanges::IRanges(c(1,201,251,1), end = c(200,250,330,100), names = letters[1:4]),
    strand = S4Vectors::Rle(BiocGenerics::strand(c("*")), 4),
    score = c(50,NA,100,30)
    )
GRange.gnr
#> GRanges object with 4 ranges and 1 metadata column:
#>     seqnames    ranges strand |     score
#>        <Rle> <IRanges>  <Rle> | <numeric>
#>   a     chr1     1-200      * |        50
#>   b     chr1   201-250      * |        NA
#>   c     chr1   251-330      * |       100
#>   d     chr2     1-100      * |        30
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
chromSize.dtf = data.frame(c("chr1","chr2"),c(350,100))
binSize.int <- 100
binnedGRanges.gnr <- BinGRanges(
    gRange.gnr = GRange.gnr,
    chromSize.dtf=chromSize.dtf,
    binSize.int=binSize.int,
    method.chr ="mean",
    variablesName.chr_vec="score",
    na.rm=TRUE
)
#> 
#>  100.00 % 
binnedGRanges.gnr
#> GRanges object with 5 ranges and 2 metadata columns:
#>       seqnames    ranges strand |     score         bin
#>          <Rle> <IRanges>  <Rle> | <numeric> <character>
#>   [1]     chr1     1-100      * |        50      chr1:1
#>   [2]     chr1   101-200      * |        50      chr1:2
#>   [3]     chr1   201-300      * |       100      chr1:3
#>   [4]     chr1   301-350      * |       100      chr1:4
#>   [5]     chr2     1-100      * |        30      chr2:1
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome