IndexFeatures

Bin multiple GRanges and summary all in one GRange. Could overlap ranges with constraints regions

IndexFeatures(
  gRange.gnr_lst = NULL,
  constraint.gnr = NULL,
  chromSize.dtf = NULL,
  binSize.int = NULL,
  method.chr = "mean",
  variablesName.chr_vec = NULL,
  cores.num = 1,
  verbose.bln = TRUE
)

Arguments

gRange.gnr_lst

<GRanges or GRangesList or list[GRanges]>: some GRanges or a list of GRanges or a GRangesList

constraint.gnr

<GRanges>: A GRange of constraint regions. If NULL chromosomes in chromSize.dtf are used (Default NULL)

chromSize.dtf

<data.frame>: A data.frame of genome where first colum correspond to the chromosomes names, and the second column correspond to the chromosomes lengths in base pairs.

binSize.int

<integer>: A number that specify the width bins.

method.chr

<character>: A string of a summary method name as 'mean', 'median', 'sum', 'max, 'min'. (Default 'mean'')

variablesName.chr_vec

<character> : A character vector that specify the metadata columns of GRanges on which apply the summary method.

cores.num

<integer> : An integer to specify the number of cores. (Default 1)

verbose.bln

<logical>: A logical value. If TRUE show the progression in console. (Default TRUE)

Value

A GRange object.

Examples

GRanges_1.gr <- GenomicRanges::GRanges(
    seqnames = S4Vectors::Rle( c("chr1", "chr2"), c(4, 2) ),
    ranges = IRanges::IRanges(
        start = c(5, 220, 260, 540, 50, 540),
        end =  c(80, 240, 280, 560, 150, 580)),
    strand = S4Vectors::Rle(BiocGenerics::strand(c("-","+","*","-","+","+")), c(1,1,1,1,1,1)),
    score = 1:6,
    name = letters[1:6]
)
GRanges_1.gr
#> GRanges object with 6 ranges and 2 metadata columns:
#>       seqnames    ranges strand |     score        name
#>          <Rle> <IRanges>  <Rle> | <integer> <character>
#>   [1]     chr1      5-80      - |         1           a
#>   [2]     chr1   220-240      + |         2           b
#>   [3]     chr1   260-280      * |         3           c
#>   [4]     chr1   540-560      - |         4           d
#>   [5]     chr2    50-150      + |         5           e
#>   [6]     chr2   540-580      + |         6           f
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
GRanges_2.gr <- GenomicRanges::GRanges(
    seqnames = S4Vectors::Rle( c("chr1", "chr2"), c(3, 2) ),
    ranges = IRanges::IRanges(
        start = c(140, 220, 450, 320, 360),
        end =  c(160, 280, 550, 350, 380)),
    strand = S4Vectors::Rle(BiocGenerics::strand(c("-","+","*")), c(1,2,2)),
    score = seq(10,50,10),
    name = letters[7:11]
)
GRanges_2.gr
#> GRanges object with 5 ranges and 2 metadata columns:
#>       seqnames    ranges strand |     score        name
#>          <Rle> <IRanges>  <Rle> | <numeric> <character>
#>   [1]     chr1   140-160      - |        10           g
#>   [2]     chr1   220-280      + |        20           h
#>   [3]     chr1   450-550      + |        30           i
#>   [4]     chr2   320-350      * |        40           j
#>   [5]     chr2   360-380      * |        50           k
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
GRanges_Constraint <- GenomicRanges::GRanges(
    seqnames = S4Vectors::Rle( c("chr1", "chr2"), c(2, 2) ),
    ranges = IRanges::IRanges(
        start = c(1, 251, 1, 50),
        end =  c(250, 490, 400, 190),
        names = paste0("C_",1:4)),
    strand = S4Vectors::Rle(BiocGenerics::strand(c("*")), c(4))
)
GRanges_Constraint
#> GRanges object with 4 ranges and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   C_1     chr1     1-250      *
#>   C_2     chr1   251-490      *
#>   C_3     chr2     1-400      *
#>   C_4     chr2    50-190      *
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
index.gnr = IndexFeatures(
    gRange.gnr_lst=list(GR_1 = GRanges_1.gr, GR_2 = GRanges_2.gr),
    constraint.gnr=GRanges_Constraint,
    chromSize.dtf=data.frame(
        seqnames=c("chr1", "chr2"),
        seqlengths=c(3500,2000)),
    binSize.int=100,
    method.chr ="mean",
    variablesName.chr_vec = c("score")
    )
#> 
#>  100.00 % 
#>  12.50 %   25.00 %  Resting time: 0.06  secs  37.50 %  Resting time: 0.06  secs  50.00 %  Resting time: 0.07  secs 
#>  100.00 % 
#>  66.67 %  Resting time: 0.00  secs  83.33 %  Resting time: 0.00  secs  100.00 % Resting time: 0.00  secs  50.00 %   100.00 % Resting time: 0.00  secs 
index.gnr
#> GRanges object with 10 ranges and 9 metadata columns:
#>        seqnames    ranges strand |                name         bin   constraint
#>           <Rle> <IRanges>  <Rle> |         <character> <character>  <character>
#>    [1]     chr1     1-100      * | chr1:1:Constraint_1      chr1:1 Constraint_1
#>    [2]     chr1   101-200      * | chr1:2:Constraint_1      chr1:2 Constraint_1
#>    [3]     chr1   201-300      * | chr1:3:Constraint_1      chr1:3 Constraint_1
#>    [4]     chr1   201-300      * | chr1:3:Constraint_2      chr1:3 Constraint_2
#>    [5]     chr1   401-500      * | chr1:5:Constraint_2      chr1:5 Constraint_2
#>    [6]     chr2     1-100      * | chr2:1:Constraint_3      chr2:1 Constraint_3
#>    [7]     chr2     1-100      * | chr2:1:Constraint_4      chr2:1 Constraint_4
#>    [8]     chr2   101-200      * | chr2:2:Constraint_3      chr2:2 Constraint_3
#>    [9]     chr2   101-200      * | chr2:2:Constraint_4      chr2:2 Constraint_4
#>   [10]     chr2   301-400      * | chr2:4:Constraint_3      chr2:4 Constraint_3
#>        GR_1.score GR_1.name  GR_1.bln GR_2.score GR_2.name  GR_2.bln
#>         <numeric>    <list> <numeric>  <numeric>    <list> <numeric>
#>    [1]        1.0         a         1         NA                   0
#>    [2]         NA                   0         10         g         1
#>    [3]        2.5       b,c         1         20         h         1
#>    [4]        2.5       b,c         1         20         h         1
#>    [5]         NA                   0         30         i         1
#>    [6]        5.0         e         1         NA                   0
#>    [7]        5.0         e         1         NA                   0
#>    [8]        5.0         e         1         NA                   0
#>    [9]        5.0         e         1         NA                   0
#>   [10]         NA                   0         45       j,k         1
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths